Move all conf validators into a common package
This makes it so that package imports don't determine anymore if a validation is actually run or not. Before, if some import tree changed and a `init` function isn't run in the frontend code anymore, the validator would never trigger.

Needs more cleanup as some of these require a DB or context. They probably shouldn't, and instead be site alerts? Not sure, it feels weird to make API requests in a validator.
