+++ cmd/symbols/parser/config.go
@@ -1,11 +1,9 @@
 package parser
 
 import (
-	"fmt"
 	"sync"
 
 	"github.com/sourcegraph/sourcegraph/internal/conf"
-	"github.com/sourcegraph/sourcegraph/internal/conf/conftypes"
 	"github.com/sourcegraph/sourcegraph/internal/ctags_config"
 	"github.com/sourcegraph/sourcegraph/lib/codeintel/languages"
 )
@@ -24,29 +22,6 @@ var parserConfig = ParserConfiguration{
 }
 
 func init() {
-	// Validation only: Do NOT set any values in the configuration in this function.
-	conf.ContributeValidator(func(c conftypes.SiteConfigQuerier) (problems conf.Problems) {
-		configuration := c.SiteConfig().SyntaxHighlighting
-		if configuration == nil {
-			return nil
-		}
-
-		for language, engine := range configuration.Symbols.Engine {
-			parser_engine, err := ctags_config.ParserNameToParserType(engine)
-			if err != nil {
-				return conf.NewSiteProblems(fmt.Sprintf("Not a valid Symbols.Engine: `%s`.", engine))
-			}
-
-			language = languages.NormalizeLanguage(language)
-			if !ctags_config.LanguageSupportsParserType(language, parser_engine) {
-				return conf.NewSiteProblems(fmt.Sprintf("Not a valid Symbols.Engine for language: %s `%s`.", language, engine))
-			}
-
-		}
-
-		return nil
-	})
-
 	// Update parserConfig here
 	go func() {
 		conf.Watch(func() {
