+++ enterprise/cmd/frontend/internal/auth/azureoauth/provider.go
@@ -32,13 +32,9 @@ const (
 func Init(logger log.Logger, db database.DB) {
 	const pkgName = "azureoauth"
 	logger = logger.Scoped(pkgName, "Azure DevOps OAuth config watch")
-	conf.ContributeValidator(func(cfg conftypes.SiteConfigQuerier) conf.Problems {
-		_, problems := parseConfig(logger, cfg, db)
-		return problems
-	})
 
 	go conf.Watch(func() {
-		newProviders, _ := parseConfig(logger, conf.Get(), db)
+		newProviders, _ := ParseConfig(logger, conf.Get(), db)
 		if len(newProviders) == 0 {
 			providers.Update(pkgName, nil)
 			return
@@ -63,7 +59,7 @@ type Provider struct {
 	providers.Provider
 }
 
-func parseConfig(logger log.Logger, cfg conftypes.SiteConfigQuerier, db database.DB) (ps []Provider, problems conf.Problems) {
+func ParseConfig(logger log.Logger, cfg conftypes.SiteConfigQuerier, db database.DB) (ps []Provider, problems conf.Problems) {
 	callbackURL, err := azuredevops.GetRedirectURL(cfg)
 	if err != nil {
 		problems = append(problems, conf.NewSiteProblem(err.Error()))
