+++ enterprise/cmd/frontend/internal/auth/bitbucketcloudoauth/config_test.go
@@ -185,7 +185,7 @@ func TestParseConfig(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			gotProviders, gotProblems := parseConfig(logtest.Scoped(t), tt.args.cfg, db)
+			gotProviders, gotProblems := ParseConfig(logtest.Scoped(t), tt.args.cfg, db)
 			gotConfigs := make([]oauth2.Config, len(gotProviders))
 			for k, p := range gotProviders {
 				if p, ok := p.Provider.(*oauth.Provider); ok {
