--- internal/auth/validateconf/BUILD.bazel
@@ -1,23 +1,0 @@
-load("//dev:go_defs.bzl", "go_test")
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "validateconf",
-    srcs = ["config.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/internal/auth/validateconf",
-    visibility = ["//:__subpackages__"],
-    deps = [
-        "//internal/conf",
-        "//internal/conf/conftypes",
-    ],
-)
-
-go_test(
-    name = "validateconf_test",
-    srcs = ["config_test.go"],
-    embed = [":validateconf"],
-    deps = [
-        "//internal/conf",
-        "//schema",
-    ],
-)
