+++ internal/conf/reposource/custom.go
@@ -9,19 +9,7 @@ import (
 
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/conf"
-	"github.com/sourcegraph/sourcegraph/internal/conf/conftypes"
 )
-
-func init() {
-	conf.ContributeValidator(func(c conftypes.SiteConfigQuerier) (problems conf.Problems) {
-		for _, c := range c.SiteConfig().GitCloneURLToRepositoryName {
-			if _, err := regexp.Compile(c.From); err != nil {
-				problems = append(problems, conf.NewSiteProblem(fmt.Sprintf("Not a valid regexp: %s. See the valid syntax: https://golang.org/pkg/regexp/", c.From)))
-			}
-		}
-		return
-	})
-}
 
 type cloneURLResolver struct {
 	from *regexp.Regexp
