+++ internal/licensing/licensing.go
@@ -160,7 +160,7 @@ func GetConfiguredProductLicenseInfoWithSignature() (*Info, string, error) {
 				return nil, "", err
 			}
 
-			if err = info.hasUnknownPlan(); err != nil {
+			if err = info.HasUnknownPlan(); err != nil {
 				return nil, "", err
 			}
