+++ internal/session/session.go
@@ -16,7 +16,6 @@ import (
 
 	"github.com/sourcegraph/sourcegraph/internal/actor"
 	"github.com/sourcegraph/sourcegraph/internal/conf"
-	"github.com/sourcegraph/sourcegraph/internal/conf/conftypes"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/env"
 	"github.com/sourcegraph/sourcegraph/internal/errcode"
@@ -42,23 +41,6 @@ const defaultExpiryPeriod = 90 * 24 * time.Hour
 
 // cookieName is the name of the HTTP cookie that stores the session ID.
 const cookieName = "sgs"
-
-func init() {
-	conf.ContributeValidator(func(c conftypes.SiteConfigQuerier) (problems conf.Problems) {
-		if c.SiteConfig().AuthSessionExpiry == "" {
-			return nil
-		}
-
-		d, err := time.ParseDuration(c.SiteConfig().AuthSessionExpiry)
-		if err != nil {
-			return conf.NewSiteProblems("auth.sessionExpiry does not conform to the Go time.Duration format (https://golang.org/pkg/time/#ParseDuration). The default of 90 days will be used.")
-		}
-		if d == 0 {
-			return conf.NewSiteProblems("auth.sessionExpiry should be greater than zero. The default of 90 days will be used.")
-		}
-		return nil
-	})
-}
 
 // sessionInfo is the information we store in the session. The gorilla/sessions library doesn't appear to
 // enforce the maxAge field in its session store implementations, so we include the expiry here.
