+++ internal/tracer/tracer.go
@@ -1,9 +1,7 @@
 package tracer
 
 import (
-	"fmt"
 	"sync/atomic"
-	"text/template"
 
 	"github.com/sourcegraph/log"
 	"go.opentelemetry.io/otel"
@@ -11,7 +9,6 @@ import (
 	oteltrace "go.opentelemetry.io/otel/trace"
 	"go.uber.org/automaxprocs/maxprocs"
 
-	"github.com/sourcegraph/sourcegraph/internal/conf"
 	"github.com/sourcegraph/sourcegraph/internal/conf/conftypes"
 	"github.com/sourcegraph/sourcegraph/internal/env"
 	"github.com/sourcegraph/sourcegraph/internal/hostname"
@@ -106,18 +103,6 @@ func Init(logger log.Logger, c WatchableConfigurationSource) {
 	// Initially everything is disabled since we haven't read conf yet - start a goroutine
 	// that watches for updates to configure the undelrying provider and debugMode.
 	go c.Watch(newConfWatcher(logger, c, provider, newOtelSpanProcessor, debugMode))
-
-	// Contribute validation for tracing package
-	conf.ContributeWarning(func(c conftypes.SiteConfigQuerier) conf.Problems {
-		tracing := c.SiteConfig().ObservabilityTracing
-		if tracing == nil || tracing.UrlTemplate == "" {
-			return nil
-		}
-		if _, err := template.New("").Parse(tracing.UrlTemplate); err != nil {
-			return conf.NewSiteProblems(fmt.Sprintf("observability.tracing.traceURL is not a valid template: %s", err.Error()))
-		}
-		return nil
-	})
 }
 
 func newTracer(logger log.Logger, provider *oteltracesdk.TracerProvider, debugMode *atomic.Bool) oteltrace.TracerProvider {
